% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{find_mangoro_bin}
\alias{find_mangoro_bin}
\title{Find the path to a mangoro Go binary}
\usage{
find_mangoro_bin(name)
}
\arguments{
\item{name}{Name of the binary (e.g. "echo")}
}
\value{
Full path to the binary in the installed package
}
\description{
Find the path to a mangoro Go binary
}
