% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{find_mangoro_vendor}
\alias{find_mangoro_vendor}
\title{Find the path to the mangoro vendor directory}
\usage{
find_mangoro_vendor()
}
\value{
Path to the vendor directory (inst/go/vendor)
}
\description{
Find the path to the mangoro vendor directory
}
