% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{get_arrow_go_version}
\alias{get_arrow_go_version}
\title{Get the version of vendored Arrow Go using Go tooling (no jsonlite)}
\usage{
get_arrow_go_version()
}
\value{
The version string of github.com/apache/arrow/go/v18 in the vendor go.mod
}
\description{
Get the version of vendored Arrow Go using Go tooling (no jsonlite)
}
