% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{get_mangos_version}
\alias{get_mangos_version}
\title{Get the version of vendored mangos using Go tooling (no jsonlite)}
\usage{
get_mangos_version()
}
\value{
The version string of go.nanomsg.org/mangos/v3 in the vendor go.mod
}
\description{
Get the version of vendored mangos using Go tooling (no jsonlite)
}
