% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{mangoro_go_build}
\alias{mangoro_go_build}
\title{Compile a Go source file using the vendored dependencies}
\usage{
mangoro_go_build(src, out, gomaxprocs = 1, gocache = NULL, ...)
}
\arguments{
\item{src}{Path to the Go source file}

\item{out}{Path to the output binary}

\item{gomaxprocs}{Number of threads for Go build (sets GOMAXPROCS env variable)}

\item{gocache}{Path to Go build cache directory. If NULL (default), uses a temporary directory to comply with CRAN policy. Set to NA to use the default Go cache location.}

\item{...}{Additional arguments to pass to Go build}
}
\value{
Path to the compiled binary
}
\description{
Compile a Go source file using the vendored dependencies
}
