% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_http_status}
\alias{mangoro_http_status}
\title{Get HTTP server status via RPC}
\usage{
mangoro_http_status(sock)
}
\arguments{
\item{sock}{A nanonext socket connected to the HTTP server controller}
}
\value{
List with status and message
}
\description{
Get HTTP server status via RPC
}
