% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_http_stop}
\alias{mangoro_http_stop}
\title{Stop the HTTP file server via RPC}
\usage{
mangoro_http_stop(sock)
}
\arguments{
\item{sock}{A nanonext socket connected to the HTTP server controller}
}
\value{
List with status and message
}
\description{
Stop the HTTP file server via RPC
}
