% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_pack_int32}
\alias{mangoro_pack_int32}
\title{Pack a 32-bit integer to raw bytes (big-endian)}
\usage{
mangoro_pack_int32(x)
}
\arguments{
\item{x}{An integer value}
}
\value{
A raw vector of length 4
}
\description{
Pack a 32-bit integer to raw bytes (big-endian)
}
