% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_rpc_call}
\alias{mangoro_rpc_call}
\title{Call a remote function via RPC}
\usage{
mangoro_rpc_call(sock, func_name, data)
}
\arguments{
\item{sock}{A nanonext socket connected to the RPC server}

\item{func_name}{Name of the function to call}

\item{data}{Data frame or Arrow stream to send as arguments}
}
\value{
The result from nanoarrow::read_nanoarrow (typically a nanoarrow_array_stream)
}
\description{
Call a remote function via RPC
}
