% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_rpc_call_message}
\alias{mangoro_rpc_call_message}
\title{Create an RPC function call message}
\usage{
mangoro_rpc_call_message(func_name, data)
}
\arguments{
\item{func_name}{Name of the function to call}

\item{data}{Data frame or Arrow stream to send as arguments}
}
\value{
A raw vector containing the RPC call message
}
\description{
Create an RPC function call message
}
