% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_rpc_get_manifest}
\alias{mangoro_rpc_get_manifest}
\title{Get the manifest of registered functions from an RPC server}
\usage{
mangoro_rpc_get_manifest(sock)
}
\arguments{
\item{sock}{A nanonext socket connected to the RPC server}
}
\value{
A list of function signatures
}
\description{
Get the manifest of registered functions from an RPC server
}
