% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_rpc_parse_response}
\alias{mangoro_rpc_parse_response}
\title{Parse an RPC response message}
\usage{
mangoro_rpc_parse_response(response)
}
\arguments{
\item{response}{Raw vector containing the RPC response}
}
\value{
A list with components: type, func_name, error_msg, data
}
\description{
Parse an RPC response message
}
