% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_rpc_send}
\alias{mangoro_rpc_send}
\title{Send a message with retries}
\usage{
mangoro_rpc_send(sock, msg, max_attempts = 20)
}
\arguments{
\item{sock}{A nanonext socket}

\item{msg}{Message to send (raw vector)}

\item{max_attempts}{Maximum number of retry attempts (default 20)}
}
\value{
The result from nanonext::send
}
\description{
Send a message with retries
}
