% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{mangoro_unpack_int32}
\alias{mangoro_unpack_int32}
\title{Unpack a 32-bit integer from raw bytes (big-endian)}
\usage{
mangoro_unpack_int32(bytes)
}
\arguments{
\item{bytes}{A raw vector of length 4}
}
\value{
An integer value
}
\description{
Unpack a 32-bit integer from raw bytes (big-endian)
}
