% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{rpc_call}
\alias{rpc_call}
\title{Call a remote function via RPC}
\usage{
rpc_call(sock, func_name, data)
}
\arguments{
\item{sock}{A nanonext socket connected to the RPC server}

\item{func_name}{Name of the function to call}

\item{data}{Data frame or Arrow stream to send as arguments}
}
\value{
A data frame with the result
}
\description{
Call a remote function via RPC
}
