% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{rpc_manifest_request}
\alias{rpc_manifest_request}
\title{Create an RPC manifest request message}
\usage{
rpc_manifest_request()
}
\value{
A raw vector containing the manifest request
}
\description{
Create an RPC manifest request message
}
