% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-utils.R
\name{rpc_recv}
\alias{rpc_recv}
\title{Receive a message with retries}
\usage{
rpc_recv(sock, max_attempts = 20)
}
\arguments{
\item{sock}{A nanonext socket}

\item{max_attempts}{Maximum number of retry attempts (default 20)}
}
\value{
The received message as a raw vector
}
\description{
Receive a message with retries
}
