% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoro-bin.R
\name{run_mangoro_bin}
\alias{run_mangoro_bin}
\title{Run a mangoro Go binary with arguments}
\usage{
run_mangoro_bin(name, args = character(), ...)
}
\arguments{
\item{name}{Name of the binary (e.g. "echo")}

\item{args}{Arguments to pass to the binary}

\item{...}{Additional arguments passed to processx::process$new}
}
\value{
A processx process object
}
\description{
Run a mangoro Go binary with arguments
}
